==> FigureYa177RNAvelocity/scripts/s2_runCCA.Rmd <==
---
title: "Run CCA"
author: "Jarning"
date: "2020/4/25"
output: html_document
---
# Academic Citation
If you use this code in your work or research, we kindly request that you cite our publication:

Xiaofan Lu, et al. (2025). FigureYa: A Standardized Visualization Framework for Enhancing Biomedical Data Interpretation and Research Efficiency. iMetaMed. https://doi.org/10.1002/imm3.70005


```{r}
source("install_dependencies.R")
knitr::opts_chunk$set(eval = FALSE, warning = FALSE, message = FALSE)
```

```{r, eval=TRUE}
oldpath <- .libPaths()[1]
newpath <- "/home/biostacs/R/backup/env_31723062"
```

加载包
```{r, eval=TRUE}
library(Seurat, lib.loc = newpath)
packageVersion("Seurat")
```

> For some specifc analyses below, we partly applied more restricted values. Gene expression (in UMI) was log normalized and scaled. Variable genes for each of the 6 samples were detected by the Seurat “FindVariableGenes” function using default options. Clustering was done using only variable genes. To obtain a global set of variable genes, we combined the set of variable genes from all 6 samples using the union of the top 1000 variable genes in each one.

导入第一部分的结果
```{r}
if (!exists("expr_mm")) {
  expr_mm <- readRDS("tmp/s1-expr_mm.rds")
}
summary(expr_mm$cells$sample_id)
```

```{r, include=FALSE, eval=TRUE}
seu.CCA <- readRDS("tmp/s2-seu_CCA.rds")
```

创建Seurat对象
```{r}
seu <- CreateSeuratObject(expr_mm$mtx, meta.data = expr_mm$cells)
```

提取每个样本的Seurat对象，共有6个样本
```{r}
sample_ids <- c(paste0("old", 1:3), paste0("young", 1:3))
seu.list <- lapply(sample_ids, function(x)
  SubsetData(seu, cells.use = rownames(subset(seu@meta.data, sample_id == x)))
  )
names(seu.list) <- sample_ids
```

对每个样本都进行Normalization、Scale并计算high variable genes
```{r}
for(i in seq_along(sample_ids)) {
  seu.list[[i]] <- NormalizeData(seu.list[[i]])
  seu.list[[i]] <- ScaleData(seu.list[[i]])
  seu.list[[i]] <- FindVariableGenes(seu.list[[i]], do.plot = FALSE)
}
```

对每个样本top 1000 high variable genes取并集
```{r}
hvg.list <- lapply(seu.list, function(x) rownames(head(x@hvg.info, n=1000)))
hvg.genes <- unique(do.call(c, hvg.list))
```

run CCA on each sample
```{r}
seu.CCA <- RunMultiCCA(seu.list, genes.use = hvg.genes, num.ccs = 20)
```

清理内存
