==> FigureYa177RNAvelocity/scripts/s6_trajectory.Rmd <==
---
title: "Trajectory Analysis"
author: "Jarning"
date: "2020/4/25"
output: html_document
---
# Academic Citation
If you use this code in your work or research, we kindly request that you cite our publication:

Xiaofan Lu, et al. (2025). FigureYa: A Standardized Visualization Framework for Enhancing Biomedical Data Interpretation and Research Efficiency. iMetaMed. https://doi.org/10.1002/imm3.70005


```{r, include=FALSE}
source("install_dependencies.R")
knitr::opts_chunk$set(eval = FALSE, warning = FALSE, message = FALSE)
```

```{r, eval=TRUE}
oldpath <- .libPaths()[1]
newpath <- "/home/biostacs/R/backup/env_31723062"
```

> *Subclustering based on DEGs* <br>
  DEGs between old and young fbroblasts were used for subclustering and
detecting cell populations of fbroblasts based on the signal produced by these genes. We used Monocle2
sorting cell functions and set up the DEGs as ordering genes for reducing dimensions. Seurat objects were
converted into Monocle (v2.4) (36) objects using the “importCDS” function. Two dimension reduction
algorithms were used for graphical representation of the data, the tSNE and DDRTree. For tSNE we used
the following parameters: num_dim = 15, max_components = 2, and norm_method = log. The clustering
was performed using the following parameters: rho_threshold = 2, and delta_threshold = 5. For DDRTree
we used this as the parameter: auto_param_selection = T.
  The DDRTree graph representation helped us visualize the progress of possible cell transitions within
the cell types using aging as the sorting factor. We plotted the subclusters split by age and states split by
subcluster and age.

```{r, eval=TRUE}
library(monocle, lib.loc = newpath)
library(Seurat, lib.loc = newpath)
packageVersion("monocle")
```

导入第四部分和第五部分的结果
```{r}
if (!exists("seu.CCA")) {
  seu.CCA <- readRDS("tmp/s4.2-seu_CCA_clustering_cellAnnotation.rds")
}
res <- readRDS("tmp/s5.2-DEG_fibroblast.rds")
```

获取差异基因
```{r}
fib.DEG <- rownames(subset(res, significant_change))
```

数据格式转换<br>
Seurat obj => monocle2

> Seurat objects were converted into Monocle (v2.4) (36) objects using the “importCDS” function.
